Add proxies, improve shelly (#8)
* remove trackable

* up

* move

* up

* up

* up

* up

* up

* fix

* add proxy

* add proxy

* up

* up

* UPDATE

* update

* update

* improve typeman

* update interface

* fix test

* d

* fix test

* update

* update serializer

+++ src/HomeBlaze.Tesla/WallConnector/TeslaWallConnectorLifetime.cs
@@ -5,7 +5,7 @@ namespace HomeBlaze.Tesla.WallConnector
     public class TeslaWallConnectorLifetime
     {
         [JsonPropertyName("energy_wh")]
-        public double TotalConsumedEnergy { get; set; }
+        public decimal TotalConsumedEnergy { get; set; }
 
         [JsonPropertyName("uptime_s")]
         public long UptimeSeconds { get; set; }
